requirements.txt
requests
python-dotenv
aiohttp
pandas
numpy
scikit-learn
aiohttp
requests
matplotlib
seaborn
umap-learn
kneed